requirements.txt
# requirements.txt
ollama
gradio
python-dotenv